refactor: new folder utils in datatable
